# Conflicts:
#	kadena2.0/CMS/js/app.min.js
#	kadena2.0/CMS/js/common.min.js
#	kadena2.0/Kadena2.0.BusinessLogic/Services/TemplateService.cs
#	kadena2.0/Kadena2.0.Dto/TemplatedProduct/MicroserviceResponses/TemplateServiceDocumentResponse.cs

+++ kadena2.0/Kadena2.0.Dto/CreditCard/AdditionalAmountsDto.cs
@@ -2,8 +2,8 @@
 {
     public class AdditionalAmountsDto
     {
-        public int ShippingAmount { get; set; }
-        public int SalesTaxAmount { get; set; }
-        public int ShippingTax { get; set; }
+        public decimal ShippingAmount { get; set; }
+        public decimal SalesTaxAmount { get; set; }
+        public decimal ShippingTax { get; set; }
     }
 }
